services/ops_agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
requests>=2.32
prometheus-client>=0.21
opentelemetry-api>=1.27
opentelemetry-sdk>=1.27
opentelemetry-exporter-otlp>=1.27
PyGithub>=2.5.0

